# Various performance improvements + code cleanup