--- a/src/ReverseDiffPrototype.jl
+++ b/src/ReverseDiffPrototype.jl
@@ -37,7 +37,19 @@ end
 # ...and the rest of the code #
 ###############################
 
-include("optimize.jl")
+# Not all operations will be valid over all of these types, but that's okay; such cases
+# will simply error when they hit the original operation in the overloaded definition.
+const ARRAY_TYPES = (:AbstractArray, :AbstractVector, :AbstractMatrix, :Array, :Vector, :Matrix)
+const REAL_TYPES = (:Bool, :Integer, :Rational, :Real, :Dual)
+
+const FORWARD_UNARY_SCALAR_FUNCS = (ForwardDiff.AUTO_DEFINED_UNARY_FUNCS..., :-, :abs, :conj)
+const FORWARD_BINARY_SCALAR_FUNCS = (:*, :/, :+, :-, :^, :atan2)
+const SKIP_BINARY_SCALAR_FUNCS = (:<, :>, :(==), :(!=), :(<=), :(>=))
+
+include("optimizations/macros.jl")
+include("optimizations/scalars.jl")
+include("optimizations/arrays.jl")
+include("optimizations/elementwise.jl")
 include("backprop.jl")
 include("api.jl")