This PR...

- ...further breaks up the code for readability
- ...allows functions to cache arbitrary data in recorded nodes during the forward pass. This state can then be utilized in the reverse pass, resulting in fewer temporaries for our array optimizations.
- ...removes my handrolled `Trace` type and replaces it with a `typealias` for `Vector{TraceNode}`. At some point in development, I guess my `Trace` implementation stopped being faster than just using a `Vector`. It would be interesting to figure out why that happened.
- ...removes the `unseed!` that was performed during the reverse pass, instead relying on `jacobian` to call `unseed!` when appropriate. This makes `gradient` more efficient.
- ...makes `TraceReal` a mutable type, and removes the use of `Ref` storage. According to our benchmarks, this results in less memory usage. It also brings us slightly closer to performing an in-place forward pass over an existing tape using new values.
- ...improves our optimization for elementwise functions. Notably, our `broadcast` optimization now works for arrays of differing dimensions and for scalars + arrays.
- Removes unused, half-implemented functionality for ternary functions. I'd rather go all-in when we support those, and I think it'll be easier to do that from a clean slate.


Here are the benchmarks after this PR:

```
julia> include("benchmark/benchmark.jl")
benchmarking ∇rosenbrock(x) on 100000 elements...
  0.004367 seconds (200.01 k allocations: 10.682 MB)
  0.002171 seconds (100.00 k allocations: 3.052 MB)
  0.007725 seconds (300.01 k allocations: 13.733 MB)
done.
benchmarking ∇ackley(x) on 100000 elements...
  0.041642 seconds (1.20 M allocations: 39.674 MB, 43.29% gc time)
  0.008778 seconds
  0.055070 seconds (1.20 M allocations: 39.674 MB, 37.14% gc time)
done.
benchmarking ∇#3(x) on 20100 elements...
  0.001612 seconds (50.35 k allocations: 2.683 MB)
  0.000760 seconds (12 allocations: 391.078 KB)
  0.002451 seconds (50.36 k allocations: 3.065 MB)
done.
benchmarking neural_net_grads...
  0.000197 seconds (3.81 k allocations: 177.969 KB)
done
```

Here are the benchmarks on master:

```
julia> include("benchmark/benchmark.jl")
benchmarking ∇rosenbrock(x) on 100000 elements...
  0.005866 seconds (200.03 k allocations: 15.260 MB)
  0.003507 seconds (100.00 k allocations: 3.052 MB)
  0.009331 seconds (300.03 k allocations: 18.312 MB)
done.
benchmarking ∇ackley(x) on 100000 elements...
  0.086067 seconds (2.20 M allocations: 62.562 MB, 66.24% gc time)
  0.013219 seconds
  0.097000 seconds (2.20 M allocations: 62.562 MB, 58.93% gc time)
done.
benchmarking ∇#3(x) on 20100 elements...
  0.002025 seconds (100.55 k allocations: 3.603 MB)
  0.001385 seconds (12 allocations: 391.078 KB)
  0.003657 seconds (100.56 k allocations: 3.985 MB)
done.
benchmarking neural_net_grads...
  0.000190 seconds (6.90 k allocations: 206.156 KB)
done
```

So we've got some modest improvements here.